Create Data Wrangling - Data Standardization
This file contains code for performing data standardization, a preprocessing technique used to scale numerical data to have a mean of 0 and a standard deviation of 1. Standardization is essential for many machine learning algorithms to ensure that features are on the same scale, improving model performance and convergence during training.