Create Data Analysis - Correlation
Correlation in Data Analysis
Correlation is a statistical measure that describes the strength and direction of the relationship between two or more variables. In data analysis, correlation helps to understand how changes in one variable may be associated with changes in another, providing valuable insights for predictions, trends, and patterns.

Types of Correlation
1. Positive Correlation: When two variables move in the same direction, meaning as one variable increases, the other also increases. For example, in retail, higher advertising spending may lead to increased sales.

2. Negative Correlation: When two variables move in opposite directions, meaning as one variable increases, the other decreases. An example might be the inverse relationship between the number of competitors in a market and the price of a product.

3. Zero (No) Correlation: When there is no discernible relationship between two variables, and the changes in one variable do not predict changes in another.

Measuring Correlation
In Python, correlation is typically measured using the Pearson correlation coefficient, denoted by r. The coefficient ranges from -1 to 1, where:
r = 1 indicates a perfect positive correlation.
r = −1 indicates a perfect negative correlation.
r = 0 indicates no correlation.

One of the common Python Tools for Correlation Analysis is Corr(), with this method we can calculate the correlation between variables of type "int64" or "float64".