Create Data Analysis - Descriptive Statistics
Before doing the data analysis, it's important to first explore your data before you spend time building complicated models. One easy way to do so is to calculate some descriptive statistics for your data. Descriptive statistical analysis helps to describe the basic features of a dataset and obtains a summary of the sample and measures of the data.

One way to do this is by using the describe function in pandas. Using the describe function and applying it to the data frame, will automatically compute the basic statistics for all numerical variables.
It shows the mean, the total number of data points, the standard deviation, the quartiles, and the extreme values.

note: any NULL values are automatically skipped in these statistics.

This function will give a clearer idea of the distribution of different variables.