Create Data Analysis - Create Box Plot
Box plots are a great way to visualize numeric data since they can visualize the various distributions of the data.

The main features that the box plot shows are the median of the data which represents where the middle data point is, the upper quartile shows where the 75th percentile is, and the lower quartile shows where the 25th percentile is.

The data between the upper and lower quartiles represents the interquartile range. Next, they have the lower and upper extremes. 
These are calculated as 1.5 times the inter-quartile range above the 75th percentile and 1.5 times the IQR below the 25th percentile.

Finally, box plots also display outliers as individual dots that occur outside the upper and lower extremes. With box plots, we can easily spot outliers and also see the distribution and skewness of the data.Box plots make it easy to compare between groups.

In this example, using a boxplot in the Seaborn Library, we can see the difference between houses with a waterfront view or without a waterfront over-price feature.