Create Data Analyst - Scatter Plot
This scatter plot represents the relationship between 2 variables, the predictor variable, and the target variable. The predictor variable is the variable that you are using to predict an outcome while the target variable is the variable that you are trying to predict. Each point on the plot corresponds to an observation, with the predictor variable plotted on the horizontal axis (X-axis) and the target variable on the vertical axis (Y-axis).

The scatter plot is useful for identifying potential trends or correlations between these two variables. A positive/negative correlation is observed, indicating that as the predictor variable increases, the outcome variable increases/decreases.

For example, we want to understand the relationship between engine size and price. Could engine size possibly predict the price of a car? In this case, our predictor variable is the engine size and the target variable is the price.

To create the scatter plot, we are using the Matplotlib function scatter, taking in x and y variables.
Note:
It is always important to label your axes and write a general plot title so that you know what you're looking at. 
​